Use std::size_t for grid and partition counts in the examples
The 1d_stencil examples, the wave equation and the checkpoint example
read nx, nt, np and cp from the command line as std::uint64_t, then pass
them to functions and containers that take a std::size_t. On a 32 bit
target MSVC reports every one of those as C4244.

They are grid points, time steps and partitions, so std::size_t is what
they are. The program_options declarations move with them, since
variables_map::as needs the type it was stored with.

nd stays std::uint64_t in the two files that have it, because that is
what the stepper's do_work takes for it.

The two resiliency examples had the conversion the other way round: they
stored hpx::chrono::high_resolution_clock::now(), which is a timestamp,
in a std::size_t. That one becomes std::uint64_t.

file_serialization converts its stream offset difference explicitly.

This covers the 1d_stencil directory including the OpenMP, repartition
and throttle variants, which are built only under their own CMake
conditions and so report no warnings today while having the same shape.
The transpose examples are the rest of this group and follow separately.

Refs #7576

+++ examples/1d_stencil/1d_stencil_1.cpp
@@ -78,9 +78,8 @@ struct stepper
 ///////////////////////////////////////////////////////////////////////////////
 int hpx_main(hpx::program_options::variables_map& vm)
 {
-    std::uint64_t nx =
-        vm["nx"].as<std::uint64_t>();    // Number of grid points.
-    std::uint64_t nt = vm["nt"].as<std::uint64_t>();    // Number of steps.
+    std::size_t nx = vm["nx"].as<std::size_t>();    // Number of grid points.
+    std::size_t nt = vm["nt"].as<std::size_t>();    // Number of steps.
 
     if (vm.count("no-header"))
         header = false;
@@ -117,9 +116,9 @@ int main(int argc, char* argv[])
     po::options_description desc_commandline;
     desc_commandline.add_options()
         ("results", "print generated results (default: false)")
-        ("nx", po::value<std::uint64_t>()->default_value(100),
+        ("nx", po::value<std::size_t>()->default_value(100),
          "Local x dimension")
-        ("nt", po::value<std::uint64_t>()->default_value(45),
+        ("nt", po::value<std::size_t>()->default_value(45),
          "Number of time steps")
         ("k", po::value<double>(&k)->default_value(0.5),
          "Heat transfer coefficient (default: 0.5)")
