+++ examples/1d_stencil/1d_stencil_4_parallel.cpp
@@ -165,10 +165,9 @@ struct stepper
 ///////////////////////////////////////////////////////////////////////////////
 int hpx_main(hpx::program_options::variables_map& vm)
 {
-    std::uint64_t np = vm["np"].as<std::uint64_t>();    // Number of partitions.
-    std::uint64_t nx =
-        vm["nx"].as<std::uint64_t>();    // Number of grid points.
-    std::uint64_t nt = vm["nt"].as<std::uint64_t>();    // Number of steps.
+    std::size_t np = vm["np"].as<std::size_t>();    // Number of partitions.
+    std::size_t nx = vm["nx"].as<std::size_t>();    // Number of grid points.
+    std::size_t nt = vm["nt"].as<std::size_t>();    // Number of steps.
 
     if (vm.count("no-header"))
         header = false;
@@ -208,11 +207,11 @@ int main(int argc, char* argv[])
     // clang-format off
     desc_commandline.add_options()
         ("results,r", "print generated results (default: false)")
-        ("nx", value<std::uint64_t>()->default_value(10),
+        ("nx", value<std::size_t>()->default_value(10),
          "Local x dimension (of each partition)")
-        ("nt", value<std::uint64_t>()->default_value(45),
+        ("nt", value<std::size_t>()->default_value(45),
          "Number of time steps")
-        ("np", value<std::uint64_t>()->default_value(10),
+        ("np", value<std::size_t>()->default_value(10),
          "Number of partitions")
         ("k", value<double>(&k)->default_value(0.5),
          "Heat transfer coefficient (default: 0.5)")
